# apps/api/tests/fixtures/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/products/mine": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "List Mine",
        "operationId": "list_mine_products_mine_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductInfoResponseDto"
                  },
                  "title": "Response List Mine Products Mine Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/products": {
      "post": {
        "tags": [
          "products"
        ],
        "summary": "Create",
        "operationId": "create_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWriteRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductInfoResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_id}/draft": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Get Draft",
        "operationId": "get_draft_products__product_id__draft_get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductInfoResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "products"
        ],
        "summary": "Update",
        "operationId": "update_products__product_id__draft_put",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductWriteRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductInfoResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "products"
        ],
        "summary": "Delete",
        "operationId": "delete_products__product_id__draft_delete",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_id}/composition": {
      "put": {
        "tags": [
          "products"
        ],
        "summary": "Put Composition",
        "operationId": "put_composition_products__product_id__composition_put",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompositionRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompositionResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Read Composition",
        "operationId": "read_composition_products__product_id__composition_get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompositionResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_id}/settings": {
      "put": {
        "tags": [
          "products"
        ],
        "summary": "Put Settings",
        "operationId": "put_settings_products__product_id__settings_put",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_id}/releases": {
      "post": {
        "tags": [
          "products"
        ],
        "summary": "Publish",
        "operationId": "publish_products__product_id__releases_post",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseInfoResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_id}/releases/{snapshot_id}/note": {
      "patch": {
        "tags": [
          "products"
        ],
        "summary": "Correct Note",
        "operationId": "correct_note_products__product_id__releases__snapshot_id__note_patch",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          },
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Snapshot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseInfoResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_id}": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Published",
        "operationId": "published_products__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishedProductInfoResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_id}/statistics": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Statistics",
        "operationId": "statistics_products__product_id__statistics_get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date To"
            }
          },
          {
            "name": "snapshot_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsInfoResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_id}/releases/{snapshot_id}/availability": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Version Availability",
        "operationId": "version_availability_products__product_id__releases__snapshot_id__availability_get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          },
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionAvailabilityInfoResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompositionRequestDto": {
        "properties": {
          "characters": {
            "items": {
              "$ref": "#/components/schemas/CharacterSelectionRequestDto"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Characters",
            "default": []
          },
          "lorebooks": {
            "items": {
              "$ref": "#/components/schemas/LorebookSelectionRequestDto"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Lorebooks",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CompositionRequestDto"
      },
      "CompositionResponseDto": {
        "properties": {
          "characters": {
            "items": {
              "$ref": "#/components/schemas/CharacterSelectionResponseDto"
            },
            "type": "array",
            "title": "Characters",
            "default": []
          },
          "lorebooks": {
            "items": {
              "$ref": "#/components/schemas/LorebookSelectionResponseDto"
            },
            "type": "array",
            "title": "Lorebooks",
            "default": []
          }
        },
        "type": "object",
        "title": "CompositionResponseDto"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProductInfoResponseDto": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "opening_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opening Message"
          },
          "visibility": {
            "type": "string",
            "title": "Visibility"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "opening_message",
          "visibility",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ProductInfoResponseDto"
      },
      "ProductWriteRequestDto": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "opening_message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Opening Message"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "public",
              "unlisted"
            ],
            "title": "Visibility",
            "default": "private"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ProductWriteRequestDto"
      },
...
      "ReleaseInfoResponseDto": {
        "properties": {
          "snapshot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Snapshot Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "change_kind": {
            "type": "string",
            "title": "Change Kind"
          },
          "update_policy": {
            "type": "string",
            "title": "Update Policy"
          }
        },
        "type": "object",
        "required": [
          "snapshot_id",
          "version",
          "summary",
          "body",
          "change_kind",
          "update_policy"
        ],
        "title": "ReleaseInfoResponseDto"
      },
      "ReleaseRequestDto": {
        "properties": {
          "summary": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Summary"
          },
          "body": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Body"
          },
          "auto_apply_media": {
            "type": "boolean",
            "title": "Auto Apply Media",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary",
          "body"
        ],
        "title": "ReleaseRequestDto"
      },
...
      "SettingsRequestDto": {
        "properties": {
          "model_id": {
            "type": "string",
            "format": "uuid",
            "title": "Model Id"
          },
          "reasoning_effort": {
            "type": "string",
            "title": "Reasoning Effort"
          },
          "start_entry_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Start Entry Ids"
          },
          "replacement_scope": {
            "type": "string",
            "enum": [
              "same_family",
              "same_provider",
              "allowlist"
            ],
            "title": "Replacement Scope",
            "default": "same_family"
          },
          "replacement_model_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Replacement Model Ids",
            "default": []
          },
          "unavailable_policy": {
            "type": "string",
            "enum": [
              "pause",
              "use_original_until_shutdown"
            ],
            "title": "Unavailable Policy",
            "default": "pause"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "model_id",
          "reasoning_effort",
          "start_entry_ids"
        ],
        "title": "SettingsRequestDto"
      },
      "SettingsResponseDto": {
        "properties": {
          "model_id": {
            "type": "string",
            "format": "uuid",
            "title": "Model Id"
          },
          "reasoning_effort": {
            "type": "string",
            "title": "Reasoning Effort"
          },
          "start_entry_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Start Entry Ids"
          },
          "replacement_scope": {
            "type": "string",
            "enum": [
              "same_family",
              "same_provider",
              "allowlist"
            ],
            "title": "Replacement Scope",
            "default": "same_family"
          },
          "replacement_model_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Replacement Model Ids",
            "default": []
          },
          "unavailable_policy": {
            "type": "string",
            "enum": [
              "pause",
              "use_original_until_shutdown"
            ],
            "title": "Unavailable Policy",
            "default": "pause"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "reasoning_effort",
          "start_entry_ids"
        ],
        "title": "SettingsResponseDto"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}